+++ src/main/kotlin/com/ecwid/apiclient/v3/dto/category/request/CategoryImageDeleteRequest.kt
@@ -4,7 +4,7 @@ import com.ecwid.apiclient.v3.dto.ApiRequest
 import com.ecwid.apiclient.v3.impl.RequestInfo
 
 data class CategoryImageDeleteRequest(
-	val categoryId: Int = 0
+	val categoryId: Long = 0L
 ) : ApiRequest {
 	override fun toRequestInfo() = RequestInfo.createDeleteRequest(
 		pathSegments = listOf(
